[libc++] Re-add a script to trigger the libc++ Buildkite pipeline (#146730)
This was removed in 99cdc26c94 when the LLVM monorepo Buildkite setup
was removed. Libc++ previously relied on the LLVM monorepo pipeline to
trigger its own pipeline when needed. Since there is no LLVM monorepo
pipeline anymore, we must trigger the libc++ Buildkite pipeline on every
pull request and use this script to determine which jobs to actually
start based on the files that are touched by the PR.

NOKEYCHECK=True
GitOrigin-RevId: 05ebb363c901cf2c153a5c90293b202cc4540887