[libc++] Avoid including vector in <functional> (#144310)
`vector` has been used in a very simple way in `boyer_moore_searcher`.
We can instead just use `unique_ptr<T[]>`, which is a lot simpler,
allowing us to drop the `vector` dependency while not losing any
expressiveness in the code. As a nice side effect, this also reduces the
time it takes to instantiate the `boyer_moore_searcher` constructor from
26ms to 22ms on my machine.

NOKEYCHECK=True
GitOrigin-RevId: 3ee0f97b950a550ef14e3adbdf45f507273f2190