[libcxx] Bump Container Runner Version (#147831)
This patch bumps the runner version from v3.222.0 to v3.226.0 as
v3.222.0 is too old at this point to connect to Github. This is needed
for the new premerge system given we are directly using this container.
This did not impact the existing libc++ CI as the runner was contained
in a separate container image.

NOKEYCHECK=True
GitOrigin-RevId: b12fcff4ffd74a37eea7877ff42d0990c07a19ab